feat: Making incomplete add, delete buttons
