+++ src/InputTodo.jsx
@@ -1,7 +1,13 @@
 import React from 'react';
 
-export default function InputTodo({ onChange }) {
+export default function InputTodo({ state, setState }) {
+  function handleInputTodo(e) {
+    setState({
+      ...state,
+      inputText: e.target.value,
+    });
+  }
   return (
-    <input placeholder="할 일을 입력해주세요" onChange={(e) => onChange(e)} />
+    <input placeholder="할 일을 입력해주세요" onChange={(e) => handleInputTodo(e)} />
   );
 }
