--- src/Title.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export default function Title() {
-  return (
-    <h1>To-do</h1>
-  );
-}
